perf(templating): inline injectable preparation
chore(tests): merge all di resolution tests into 1 file
